+++ tests/bootstrap.php
@@ -14,7 +14,7 @@ if (!$_tests_dir) {
 
 // Forward custom PHPUnit Polyfills configuration to PHPUnit bootstrap file.
 $_phpunit_polyfills_path = getenv('WP_TESTS_PHPUNIT_POLYFILLS_PATH');
-if (false !== $_phpunit_polyfills_path) {
+if ($_phpunit_polyfills_path !== false) {
     define('WP_TESTS_PHPUNIT_POLYFILLS_PATH', $_phpunit_polyfills_path);
 }